First init implementation
