+++ internal/engine/ops/audit_test.go
@@ -1,4 +1,4 @@
-package engine
+package ops
 
 import (
 	"context"
@@ -8,44 +8,8 @@ import (
 	"strings"
 	"testing"
 
-	"gaal/internal/config"
+	"gaal/internal/engine/render"
 )
-
-// ── Audit — empty environment ────────────────────────────────────────────────
-
-func TestAudit_EmptyEnv_Table(t *testing.T) {
-	e := NewWithOptions(nil_config(t), Options{WorkDir: t.TempDir()})
-	out := captureStdout(t, func() {
-		if err := e.Audit(context.Background(), FormatTable); err != nil {
-			t.Errorf("Audit table (empty env): %v", err)
-		}
-	})
-	if !strings.Contains(out, "Discovered Skills") {
-		t.Errorf("output missing 'Discovered Skills' section:\n%s", out)
-	}
-	if !strings.Contains(out, "Discovered MCP Servers") {
-		t.Errorf("output missing 'Discovered MCP Servers' section:\n%s", out)
-	}
-}
-
-func TestAudit_EmptyEnv_JSON(t *testing.T) {
-	e := NewWithOptions(nil_config(t), Options{WorkDir: t.TempDir()})
-	out := captureStdout(t, func() {
-		if err := e.Audit(context.Background(), FormatJSON); err != nil {
-			t.Errorf("Audit json (empty env): %v", err)
-		}
-	})
-	var report AuditReport
-	if err := json.Unmarshal([]byte(out), &report); err != nil {
-		t.Fatalf("invalid JSON output: %v\noutput: %s", err, out)
-	}
-	// Empty env: no skills or MCPs, but keys must be present (null is ok).
-	// Just verify the JSON decoded without error and both slices are well-typed.
-	_ = report.Skills
-	_ = report.MCPs
-}
-
-// ── Audit — project skills discovery ────────────────────────────────────────
 
 // makeSkill creates <workDir>/<relDir>/<skillName>/SKILL.md with optional
 // frontmatter containing the given description.
@@ -61,15 +25,49 @@ func makeSkill(t *testing.T, workDir, relDir, skillName, desc string) {
 	}
 }
 
-func TestAudit_ProjectSkills_Table(t *testing.T) {
+// ── Audit — empty environment ────────────────────────────────────────────────
+
+func TestAudit_EmptyEnv_Table(t *testing.T) {
+	home := t.TempDir()
 	workDir := t.TempDir()
-	// .agents/skills is the canonical project_skills_dir for many agents (amp,
-	// github-copilot, etc.). A skill placed here must be discovered.
+	out := captureStdout(t, func() {
+		if err := Audit(context.Background(), home, workDir, render.FormatTable); err != nil {
+			t.Errorf("Audit table (empty env): %v", err)
+		}
+	})
+	if !strings.Contains(out, "Discovered Skills") {
+		t.Errorf("output missing 'Discovered Skills' section:\n%s", out)
+	}
+	if !strings.Contains(out, "Discovered MCP Servers") {
+		t.Errorf("output missing 'Discovered MCP Servers' section:\n%s", out)
+	}
+}
+
+func TestAudit_EmptyEnv_JSON(t *testing.T) {
+	home := t.TempDir()
+	workDir := t.TempDir()
+	out := captureStdout(t, func() {
+		if err := Audit(context.Background(), home, workDir, render.FormatJSON); err != nil {
+			t.Errorf("Audit json (empty env): %v", err)
+		}
+	})
+	var report render.AuditReport
+	if err := json.Unmarshal([]byte(out), &report); err != nil {
+		t.Fatalf("invalid JSON output: %v\noutput: %s", err, out)
+	}
+	_ = report.Skills
+	_ = report.MCPs
+}
+
+// ── Audit — project skills discovery ────────────────────────────────────────
+
+func TestAudit_ProjectSkills_Table(t *testing.T) {
+	home := t.TempDir()
+	workDir := t.TempDir()
 	makeSkill(t, workDir, ".agents/skills", "my-skill", "A test skill")
 
-	e := NewWithOptions(nil_config(t), Options{WorkDir: workDir})
 	out := captureStdout(t, func() {
-		if err := e.Audit(context.Background(), FormatTable); err != nil {
+		if err := Audit(context.Background(), home, workDir, render.FormatTable); err != nil {
 			t.Errorf("Audit table (project skills): %v", err)
 		}
 	})
@@ -79,16 +77,16 @@ func TestAudit_ProjectSkills_Table(t *testing.T) {
 }
 
 func TestAudit_ProjectSkills_JSON(t *testing.T) {
+	home := t.TempDir()
 	workDir := t.TempDir()
 	makeSkill(t, workDir, ".agents/skills", "json-skill", "JSON test skill")
 
-	e := NewWithOptions(nil_config(t), Options{WorkDir: workDir})
 	out := captureStdout(t, func() {
-		if err := e.Audit(context.Background(), FormatJSON); err != nil {
+		if err := Audit(context.Background(), home, workDir, render.FormatJSON); err != nil {
 			t.Errorf("Audit json (project skills): %v", err)
 		}
 	})
-	var report AuditReport
+	var report render.AuditReport
 	if err := json.Unmarshal([]byte(out), &report); err != nil {
 		t.Fatalf("invalid JSON: %v\noutput: %s", err, out)
 	}
@@ -109,20 +107,17 @@ func TestAudit_ProjectSkills_JSON(t *testing.T) {
 
 // ── Audit — two-pass deduplication ──────────────────────────────────────────
 
-// TestAudit_TwoPass_NoProjectDuplicate verifies that a skill placed in a
-// shared directory (e.g. .agents/skills) appears only once even though
-// multiple agents include that path in their search lists.
 func TestAudit_TwoPass_NoProjectDuplicate(t *testing.T) {
+	home := t.TempDir()
 	workDir := t.TempDir()
 	makeSkill(t, workDir, ".agents/skills", "shared-skill", "Shared skill")
 
-	e := NewWithOptions(nil_config(t), Options{WorkDir: workDir})
 	out := captureStdout(t, func() {
-		if err := e.Audit(context.Background(), FormatJSON); err != nil {
+		if err := Audit(context.Background(), home, workDir, render.FormatJSON); err != nil {
 			t.Errorf("Audit json (dedup): %v", err)
 		}
 	})
-	var report AuditReport
+	var report render.AuditReport
 	if err := json.Unmarshal([]byte(out), &report); err != nil {
 		t.Fatalf("invalid JSON: %v\noutput: %s", err, out)
 	}
@@ -137,24 +132,17 @@ func TestAudit_TwoPass_NoProjectDuplicate(t *testing.T) {
 	}
 }
 
-// TestAudit_TwoPass_CanonicalAgent verifies that a skill placed in an agent's
-// canonical project directory is attributed to that agent rather than another
-// agent that lists the same directory in its extended search list.
-//
-// .claude/skills is the canonical project_skills_dir for claude-code, but it
-// also appears in the project_skills_search of amp, github-copilot, and others.
-// After the two-pass fix, the skill must be attributed to claude-code.
 func TestAudit_TwoPass_CanonicalAgent(t *testing.T) {
+	home := t.TempDir()
 	workDir := t.TempDir()
 	makeSkill(t, workDir, ".claude/skills", "claude-skill", "Claude canonical skill")
 
-	e := NewWithOptions(nil_config(t), Options{WorkDir: workDir})
 	out := captureStdout(t, func() {
-		if err := e.Audit(context.Background(), FormatJSON); err != nil {
+		if err := Audit(context.Background(), home, workDir, render.FormatJSON); err != nil {
 			t.Errorf("Audit json (canonical agent): %v", err)
 		}
 	})
-	var report AuditReport
+	var report render.AuditReport
 	if err := json.Unmarshal([]byte(out), &report); err != nil {
 		t.Fatalf("invalid JSON: %v\noutput: %s", err, out)
 	}
@@ -167,13 +155,4 @@ func TestAudit_TwoPass_CanonicalAgent(t *testing.T) {
 		}
 	}
 	t.Error("'claude-skill' not found in JSON skills list")
-}
-
-// ── helpers ──────────────────────────────────────────────────────────────────
-
-// nil_config returns an empty *config.Config (kept as a named helper to keep
-// test bodies concise without importing config).
-func nil_config(t *testing.T) *config.Config {
-	t.Helper()
-	return &config.Config{}
 }
