+++ internal/engine/ops/init.go
@@ -1,4 +1,4 @@
-package engine
+package ops
 
 import (
 	_ "embed"
@@ -9,12 +9,12 @@ import (
 )
 
 //go:embed init_template.yaml
-var initTemplate []byte
+var InitTemplate []byte
 
 // Init writes the documented gaal.yaml skeleton to dest.
 // When force is false and dest already exists, an error is returned so the
 // caller can surface an actionable message without silently overwriting work.
-func (e *Engine) Init(dest string, force bool) error {
+func Init(dest string, force bool) error {
 	slog.Debug("init", "dest", dest, "force", force)
 
 	if !force {
@@ -25,7 +25,7 @@ func (e *Engine) Init(dest string, force bool) error {
 		}
 	}
 
-	if err := os.WriteFile(dest, initTemplate, 0o644); err != nil {
+	if err := os.WriteFile(dest, InitTemplate, 0o644); err != nil {
 		return fmt.Errorf("writing %s: %w", dest, err)
 	}
