+++ internal/engine/ops/init_template.yaml
