+++ internal/engine/render/table_test.go
@@ -1,4 +1,4 @@
-package engine
+package render
 
 import (
 	"bytes"
